branched off Linux packages to their own repo.
This is to relieve the release overhead in the main clockworkpi-linux
repository when there are updates. Most of the packages don't change
often but would still be uploaded as part of a new kernel release.

As for the Linux Kernel itself, there was surprisingly no need for
changes in the patchsets. Instead, the PKGBUILD files had to be adjusted
because they tried deleting the subdirectories "source" and "build" from
the /usr/lib/modules/.../ directory to keep them from bloating the
resulting packages but it looks like the subdirectories aren't created
anymore.